0702p0624

+++ principal_program/Core/Inc/main/adc.h
@@ -4,12 +4,12 @@
 
 typedef struct
 {
-    uint32_t sensor_track_right;
-    uint32_t sensor_track_left;
-    uint32_t sensor_node;
-    uint32_t sensor_direction;
-    uint32_t magnetic_stripe_value;         // 判斷磁條強度大小
-    uint32_t strong_magnet_value;           // 判斷強力磁鐵強度大小
+    uint16_t sensor_track_right;
+    uint16_t sensor_track_left;
+    uint16_t sensor_node;
+    uint16_t sensor_direction;
+    uint16_t magnetic_stripe_value;         // 判斷磁條強度大小
+    uint16_t strong_magnet_value;           // 判斷強力磁鐵強度大小
 } AdcHall;
 
 extern AdcHall adc_hall;
